added reader
