+++ parser.js
@@ -1,29 +1,9 @@
-const fs        = require('fs');
-
-/*** Synchronous Version ***/
-/*
-const parseSources = ({ source, static }) => {
-  return source.map((file_name) => {
-    return fs.readFileSync(file_name);
-  });
-}; */
-
-/*** Asynchronous Version ***/
-const parseSources = ({ source, static }) => {
-  let filesArray = [];
-  return new Promise((resolve) => {
-    const cont = (err, data) => {
-      console.log(data);
-      filesArray.push(data);
-      if(filesArray.length === source.length) {
-        resolve(filesArray);
-      }
-    };
-    source.forEach((file_name) => {
-      console.log(file_name);
-      fs.readFile(file_name, cont);
-    });
+const parser = readers => {
+  readers.forEach((reader) => {
+    console.log("Printing Reader in Parser\n-----------------------\n");
+    reader.printFile();
+    console.log("\n\n");
   });
 };
 
-module.exports = parseSources;
+module.exports = parser;
