+++ tests/reader/config_reader_test.js
@@ -5,7 +5,7 @@ const assert            = require('assert');
 
 const e                 = require('../../constants/errors');
 const _                 = require('../../constants/globals');
-const config_reader     = require('../../reader/config_reader');
+const config_reader     = require('../../readers/config_reader');
 
 chai.use(chai_as_promised);
 
@@ -22,10 +22,10 @@ describe('JSON Reader', function() {
         }
         fs.writeFileSync(file_name, JSON.stringify(obj), 'utf8');
         const res = await jsonReader(file_name);
-        // deeply compare objects 
+        // deeply compare objects
         chai.expect(res).to.deep.equal(obj);
     });
-    
+
     it('should throw an error for a file with invalid JSON', () => {
         // Write invalid JSON to file
         fs.writeFileSync(file_name, "This is not json");
@@ -47,7 +47,7 @@ describe('JSON Reader', function() {
 });
 
 describe('Check Config', function() {
-    const {checkConfig} = config_reader; 
+    const {checkConfig} = config_reader;
 
     it('should throw no errors for a valid object', function() {
         const obj = {
